refactor: split three files back under the 500-LoC cap
CI caught this, and it was a real miss on my part: I checked the line count of
every file I ADDED and none of the ones I grew.

  bench_run.rs   454 -> 660 -> 373   (card + record writing moved out)
  check.rs       499 -> 516 -> 494   (message rendering moved out)

`bench_run.rs` splits on the seam `bench_state_history.rs` already established:
`bench_card.rs` is the four members that render an SVG rather than drive a
benchmark, `bench_record.rs` the one that turns a finished run into the committed
artefacts a merge depends on. Exact piecewise move, no logic changed.

`check.rs` was at 499 before this branch — one line under the cap — so seventeen
lines of new signature check put it over. Two changes, in this order:

* The "why fail, not skip" rationale moved to `signing::verify_record`, beside
  the function that decides, leaving a pointer. The explanation is worth more
  next to the code it explains than next to the call.
* `summarize_paths` moved to a new `check_fmt.rs`. That block was chosen
  deliberately: `hardening_tests` pins every verdict-DECIDING symbol inside
  `BOUNDARY_FILES`, and `summarize_paths` decides nothing — it turns a list of
  paths into one readable line. Moving `record_covers` or `invalidating_paths`
  instead would have put a verdict outside the boundary, which is precisely the
  "lock whose key is kept inside it" shape `coverage.rs` exists to prevent. The
  new file's header says so, so the next person to reach for it knows why it is
  not a boundary file.

The allow-list in `file-size-cap.yml` was the other option and is the wrong one
here: its own rule asks for a rationale and a tracking issue, and both of these
had a clean split available.

900 atlas-plugin tests, 2334 spark-server, fmt and clippy clean.
