+++ crates/spark-server/src/tui/bench_state_history.rs
@@ -46,7 +46,7 @@ impl BenchState {
             return crate::tui::bench_keys::Outcome::None;
         };
         self.status =
-            match crate::cli::bench_run::render_card_for_benchmark(&run.benchmark_id, None) {
+            match crate::cli::bench_card::render_card_for_benchmark(&run.benchmark_id, None) {
                 Ok(path) => format!("card written to {}", path.display()),
                 // The common cause is a benchmark that has never been committed as a
                 // gate record — say that, rather than "failed".
